Display roll call on motion detail view
This commit adds a table with headers "YES" and "NO" to the motion detail view.
The headers also show the total number of users who have voted on each side.

Beneath the headers are the usernames of each user who voted on each side.

To implement this, a new value "votersByVote" has been added to the Motion
class. It is lazily evaluated by inverting the motion's roll call.